backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
aiofiles
pyyaml
